refactor: move update, revert to helpers
